first official release markabat

+++ ecs_vehicles/ecs_vehicles/report/fuel_stocktaking/fuel_stocktaking.py
@@ -49,7 +49,7 @@ def get_columns():
         {
             "label": _("الكمية باللتر"),
             "fieldname": "qty_litre",
-            "fieldtype": "Float",
+            "fieldtype": "Integer",
             "width": 130
         },
         {
