+++ ecs_vehicles/ecs_vehicles/report/gas_stocktaking/gas_stocktaking.py
@@ -49,7 +49,7 @@ def get_columns():
         {
             "label": _("الكمية بالمتر المكعب"),
             "fieldname": "qty_litre",
-            "fieldtype": "Float",
+            "fieldtype": "Integer",
             "width": 150
         },
         {
